requirements.txt
attrs==24.2.0
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.4.0
forex-python==1.8
fredapi==0.5.2
frozendict==2.4.6
h11==0.14.0
html5lib==1.1
idna==3.10
lxml==5.3.0
multitasking==0.0.11
numpy==2.1.3
outcome==1.3.0.post0
pandas==2.2.3
pandas-datareader==0.10.0
peewee==3.17.8
platformdirs==4.3.6
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
selenium==4.26.1
setuptools==75.5.0
simplejson==3.19.3
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
webencodings==0.5.1
websocket-client==1.8.0
wsproto==1.2.0
yfinance==0.2.49
